Condition
To write a program that reads 2 * n number of integers supplied by the user and checks if the sum of the first n numbers (left sum) is equal to the sum of the second n numbers (right sum). If there is a tie, print " Yes, sum = " + the sum; otherwise it prints " No, diff = " + the difference. The difference is calculated as a positive number (in absolute value).