Condition
Given n integers in interval [1…1000]. Of these, some percent p1 is below 200, another percent p2 is from 200 to 399, another percent p3 is from 400 to 599, another percent p4 is from 600 to 799, and the remaining p5 percent is from 800 and up. To write a program that calculates and prints the percentages p1, p2, p3, p4 and p5.
Input:
The first line of the input contains the integer n (1 ≤ n ≤ 1000) - number of numbers. On the next n lines there is one whole number in the interval [1...1000] - the numbers on which the histogram should be calculated.
Output:
Print the histogram to the console - 5 lines, each containing a number between 0% and 100%, to two decimal places, eg 25.00%, 66.67%, 57.14%.
Example:
Input:        Output:                           Input:        Output:
7               14.29%                            1                66.67%
800           28.57%                            2                0.00%
801           14.29%                            999            0.00%
250           14.29%                                              0.00%
199           28.57%                                              33.33%
399
599
799