Condition
Write a program to calculate exam grade statistics. At the beginning, the program receives the number of students who took the exam and for each student his grade. At the end, the program should print the percentage of students with a grade between 2.00 and 2.99, between 3.00 and 3.99, between 4.00 and 4.99, 5.00 or higher. Also the GPA on the exam.
Input:
A series of numbers are read from the console, each on a separate line:
• On the first line - the number of students who took the exam - an integer in the interval [1...1000]
• For each student on a separate line – the grade from the exam – a real number in the interval [2.00...6.00]
Output:
To print to the console 5 lines containing the following information:
Line 1 - "Top students: {percent of students with GPA 5.00 or higher}%"
Line 2 - "Between 4.00 and 4.99: {between 4.00 and 4.99 inclusive}%"
Line 3 - "Between 3.00 and 3.99: {between 3.00 and 3.99 inclusive}%"
Line 4 - "Fail: {less than 3.00}%"
Line 5 - "Average: {average success}"
All numbers must be formatted to the second decimal place.
Example:
Input:        Output:                                                                          Input:        Ouput:
10             Top students: 30.00%                                                     6               Top students: 33.33%
3.00          Between 4.00 and 4.99: 30.00%                                     2                Between 4.00 and 4.99: 16.67%
2.99          Between 3.00 and 3.99: 20.00%                                     3                Between 3.00 and 3.99: 16.67%
5.68          Fail: 20.00%                                                                    4                Fail: 3.33%
3.01          Average: 4.06                                                                 5                Average: 3.70
4                                                                                                     6
4                                                                                                     2.2
6.00
4.50
2.44
5