Condition
Write a program to calculate the score of the game. First you get a number that indicates how many turns the game will last. After each turn of the game you will get a new number. Points are added according to the interval in which the number falls. If the number is negative or greater than 50, then it is invalid. At the start of the game the score is 0 and points are added each turn as follows:
• From 0 to 9 - 20 % of the number
• From 10 to 19 - 30 % of the number
• From 20 to 29 - 40 % of the number
• From 30 to 39 - 50 points
• From 40 to 50 - 100 points
• Invalid number - the result is divided by 2
In addition to the result, the program must generate statistics for percentages of numbers in the given intervals.
Input:
The input is read from the console:
• First row - how many moves there will be during the game - an integer in the interval [1...100]
• For each move - the numbers to be checked in which interval are - integers in the interval [-100...100]
Output:
To print to the console 7 lines:
• 1st line: "{Final Result}"
• 2nd line: "From 0 to 9: {percent in interval}%"
• 3rd line: "From 10 to 19: {percent in interval}%"
• 4th line: "From 20 to 29: {percent in interval}%"
• 5th line: "From 30 to 39: {percent in interval}%"
• 6th line: "From 40 to 50: {percent in interval}%"
• 7th line: "Invalid numbers: {percent in interval}%"
All numbers must be formatted to the second decimal place.
Example:
Input:        Output:
10              295.80
43              From 0 to 9: 10.00%
57              From 10 to 19: 10.00%
-12            From 20 to 29: 20.00%
23             From 30 to 39: 10.00%
12             From 40 to 50: 30.00%
0               Invalid numbers: 20.00%
50
40
30
20