Condition
The SoftUni team is organizing a football tournament. First we read from the console the capacity of the stadium and the number of all fans. Then, for each fan, the sector in which it is located is read. The first team's fans are in sectors A and B, and the second team's fans are in C and D. Write a program that calculates the percentages of fans in each sector, relative to the total number of fans in the stadium, as well as the total percentage of fans for both teams. according to the capacity of the stadium. The total number of fans does NOT exceed the capacity of the stadium.
Input:
A series of numbers are read from the console, each on a separate line:
1. The capacity of the stadium – an integer in the interval [1 … 10000];
2. The number of all fans – an integer in the interval [1 … 10000].
For each fan, on a separate line read:
1. The sector it is on - text - "A", "B", "V" and "G".
Output:

To print to the console 5 lines, each containing a percentage between 0.00% and 100.00%, formatted to the second digit after the decimal point:
1. The percentage of fans in sector A
2. The percentage of fans in sector B
3. The percentage of fans in sector B
4. The percentage of fans in sector D
5. The percentage of all fans, relative to the capacity of the stadium.
Example:
Input:        Output:
93             25.00%
16             50.00%
A               6.25%
V               18.75%
G               17.20%
G
B
B
G
B
A
B
B
B
A
B
B
A