Condition
Bozidara has several houses on the Black Sea coast and wishes to green the yards of some of them, thus creating a cozy atmosphere and comfort for its guests. For this purpose she has hired a company.
Write a program that calculates the required amount that Bozidara will have to pay to the contractor of the project. The cost per square meter is 7.61 BGN including VAT. Because her yard is quite large, the contractor offers an 18% discount on the final price.
Input:
Only one line is read from the console:
1.	Square meters to be landscaped - real number in the range [0.00 ... 10000.00]
Output:
Two lines are printed on the console:
- "The final price is: {final service price} lv."
- "The discount is: {discount} lv."
Example:
Input:        Output:                                                                                      Input:       Output:
550           The final price is: 3432.11 lv.                                                      150          The final price is: 936.03 lv.
                 The discount is: 753.39 lv.                                                                          The discount is: 205.47 lv.