Condition
The school year has already started and the person in charge of the 10B class, Annie, needs to buy a certain number of packets of pens, packets of markers, and board cleaner. She is a regular customer of a bookstore, so there is a discount for her that represents some percentage of the total. Write a program that calculates how much money Annie will need to raise to pay the bill, given the following price list: 
- Pack of chemicals - £5.80. 
- Pack of markers - BGN 7.20. 
- A pack of markers - 1.20 BGN (per liter)
Input:
4 numbers are read from the console:
- Number of chemical packages - integer in the range [0...100]
- Number of marker packs - integer in the range [0...100]
- Liters of board cleaner - integer in the range [0...50]
- Percentage reduction - integer in the range [0...100]
Output:
Print on the console how much money Annie will need to pay her bill.
Example:
Input:        Output:                    Input:        Output:
2               28.5                          4               37.932
3                                                2
4                                                5
25                                              13