Condition
Rumen wanted to repaint the living room and hired craftsmen for this purpose. Write a program that calculates the cost of the renovation given the following prices:
- Protective nylon - BGN 1.50 per square meter
- Paint - BGN 14.50 per liter
- Paint thinner - BGN 5.00 per liter
Just in case, Rumen wants to add another 10% of the amount of paint and 2 sq.m. of nylon, of course and 0.40 BGN for bags to the required materials. The amount to be paid to the craftsmen for 1 hour of work is equal to 30% of the sum of all material costs.
Input:
The input is read from the console and contains exactly 4 lines:
1.	Amount of nylon needed (in square meters) - integer in the range [1... 100].
2.	Required amount of paint (in liters) - integer in the range [1...100]
3.	Quantity of thinner (in liters) - integer in the range [1...30]
4.	Hours for which the craftsmen will do the work - integer in the range [1...9]
Output:
Print one line on the console:
- "{sum of all costs}"
Example:
Input:        Output:                                     Input:                Output:
10             727.09                                       5                       286.52
11                                                               10
4                                                                 10
8                                                                 1