Condition
The restaurant opens its doors and offers several menus at discounted prices: 
- Chicken menu - BGN 10.35. 
- Menu with fish - 12.40 BGN. 
- Vegetarian menu - BGN 8.15. 
Write a program that calculates how much it will cost for a group of people to order a takeaway.
The group will also order a dessert whose cost is equal to 20% of the total bill (excluding delivery). 
The cost of delivery is £2.50 and is charged at the end.  
Input:
3 lines are read from the console:
- Number of chicken menus - integer in the range [0 ... 99]
- Number of fish menus - integer in the range [0 ... 99]
- Number of vegetarian menus - integer in the range [0 ... 99]
Output:
Print one line on the console: "{order price}"
Example:
Input:        Output:                                 Input:        Output:
2               116.2                                     9               202.72
4                                                             2
3                                                             6