Condition
For his birthday, Lubomir received an aquarium in the shape of a parallelepiped. At first we read from the console on separate lines its dimensions - length, width and height in centimeters. We have to calculate how many litres of water the aquarium will hold, knowing that a certain percentage of its capacity is taken up by sand, plants, a heater and a pump. 
One litre of water equals one cubic decimetre / 1l=1dm3/. 
Write a program that calculates the litres of water needed to fill the aquarium.
Input:
4 lines are read from the console:
1.	Length in cm - integer in the range [10 ... 500]
2.	Width in cm - integer in the range [10 ... 300]
3.	Height in cm - integer in the range [10 ... 200]
4.	Percentage - real number in the range [0.000 ... 100.000]
Output:
Print a number on the console:
- The liters of water the aquarium will collect.
Example:
Input:        Output:                                          Input:        Output:
85             248.68875                                      105            586.445475
75                                                                    77
47                                                                    89
17                                                                    18.5