Condition
Write a program that reads three fractional numbers b1, b2 and h from the console and computes the face of a trapezoid with bases b1 and b2 and height h. The formula for the face of a trapezoid is (b1 + b2) * h / 2.
The figure below shows a trapezoid with sides 8 and 13 and height 7. It has face (8 + 13) * 7 / 2 = 73.5.
The answer must be formatted to the second digit after the decimal place.
Example:
Input:        Output: 
8               73.50                              
13
7