Condition
A classroom has a rectangular size w by h meters, without columns inside. The hall is divided into two parts, left and right, with a corridor approximately in the middle. There are rows of desks in the left and right parts. At the back of the hall is a large doorway. At the front of the hall is a lectern with a podium for the lecturer. One workstation occupies 70 by 120 cm (70 by 40 cm table + 70 by 80 cm chair and passing space). The corridor is at least 100 cm wide. It has been calculated that exactly 1 workstation is lost because of the entrance door (which has an opening of 160 cm) and exactly 2 workstations are lost because of the chair (which is 160 cm by 120 cm). Write a program that enters the dimensions of the classroom and calculates the number of workstations in the classroom for the described layout 
Input:
The console reads 2 numbers, one per line: w (length in meters) and h (width in meters).
Constraints: 3 ≤ h ≤ w ≤ 100.
Output:
Print an integer on the console: the number of seats in the classroom.
Example:
Input:        Output:                           Input:        Output:
15             129                                  8.4            39
8.9                                                    5.2