Condition
George will have guests in the evening and decides to treat them with bonito, horse mackerel and mussels. So he goes to the fish market to buy a few kilos. From the console the prices in euros of the mackerel and sprat are entered. Also the quantity of bonito, horse mackerel and mussels in kilograms. How much money will he need to pay his bill if the prices on the exchange are:
- Palamod - 60% more expensive than mackerel
- Mackerel - 80% more expensive than sprat
- Mussels - BGN 7.50 per kilo
Input:
5 numbers are read from the console:
- First line - price of mackerel per kilo. Real number in the range [0.00...40.00]
- Second line - price of sprat per kilo. Real number in the range [0.00...30.00]
- Third row - kilo of bonito. Real number in the range [0.00...50.00]
- Fourth row - kilograms of horse mackerel. Real number in the range [0.00... 70.00]
- Fifth row - kilograms of mussels. Integer in the range [0 ... 100]
Output:
Print a floating-point number on the console: how much money George will need to pay his bill. The number must be formatted to the second decimal place (1.2457 -> 1.25).
Example:
Input:        Output:                                 Input:        Output:
6.90           42.96                                    5.55           113.82
4.20                                                        3.57
1.5                                                          4.3
2.5                                                          3.6
1                                                             7