Condition
Write a program to calculate how many liters of paint are needed to paint a house. Using green paint for the walls and red paint for the roof. The consumption of the green paint is 1 liter for 3.4 m2 and the consumption of the red paint is 1 liter for 4.3 m2.
The walls have the following dimensions:
- The front and back walls are squares with side "x".
o the front wall has a rectangular door 1.2m wide and 2m high
- The side walls are rectangles with sides "x" and "y"
o on both side walls there is one square window with 1.5m side
The roof has the following dimensions:
- Two rectangles with sides "x" and "y"
- Two equilateral triangles with side "x" and height "h"
You need to calculate the area of all the sides and the area of the roof to
find how many liters of each paint will be needed.
Input:
From the console, 3 lines are read:
1.	x - the height of the house - a real number in the range [2...100]
2.	y - the length of the side wall - a real number in the interval [2...100]
3. h - the height of the triangular wall of the roof - a real number in the interval [2...100]
Output:
Print two numbers on the console each on a new line:
- Liters of green paint
- Liters of red paint
Formatted to the second decimal place.
Example:
Input:        Output:                                 Input:        Output:
6               54.44                                    10.25          152.93
10             35.16                                    15.45          94.82
5.2                                                         8.88