Condition
Write a program that guesses whether it is hot or cold outside. A line of text is read from the console that tells you what the weather is. When "sunny" is typed, "It's warm outside!" should be printed. In all other cases "It's cold outside!" should be printed. 
Example:
Input:        Output:                                                      Input:        Output:                                            Input:        Output:
sunny        It's warm outside!                                     cloudy       It's cold outside!                              snowy       It's cold outside!