Condition
Write a program in which the user enters the type and dimensions of a geometric figure and calculates its face. There are four types of shapes: square, rectangle, circle and triangle. The first line of the input reads the type of shape (text with the following options: square, rectangle, circle or triangle). 
- If the shape is a square: the next line reads a fractional number - the length of its side
- If the shape is a rectangle: the next two lines read two fractional numbers - the lengths of its sides
- If the figure is a circle: the next line reads one fractional number - the radius of the circle
- If the figure is a triangle: the next two lines read two fractional numbers - the length of its side and the length of the height to it
Round the result to 3 decimal places. 
Example:
Input:              Output:                                     Input:           Output:
rectangle        17.500                                       triangle        45.000
7                                                                       4.5
2.5                                                                    20