Condition
An integer is given - an initial number of points. Bonus points are added to it according to the rules described below. Write a program that calculates the bonus points the number receives and the total number of points (the number + the bonus).
- If the number is up to and including 100, the bonus points are 5.
- If the number is greater than 100, the bonus points are 20% of the number.
- If the number is greater than 1000, the bonus points are 10% of the number.

- Additional bonus points (awarded separately from previous bonus points):
o For an even number + 1 point.
o For a number ending in 5 + 2 pts.
Example:
Input:        Output:                              Input:        Output:
20             6.0                                      175            37.0
                 26.0                                                      212.0