Condition
Write a program that reads the hour and minutes of a 24-hour day entered by the user and calculates what time it will be in 15 minutes. Print the result in hours:minutes format. The hours are always between 0 and 23 and the minutes are always between 0 and 59. The hours are written in one or two digits. Minutes are always written in two digits, with a leading zero where necessary. 
Example:
Input:        Output:                                      Input:        Output:
1               2:01                                            23             0:14
46                                                                59