Condition
Petya has a toy shop. She receives a big order that she has to fulfill. With the money she will earn she wants to go on a trip. 
Toy prices:
- Puzzle - BGN 2.60.
- Talking doll - BGN 3.
- Teddy bear - BGN 4.10.
- Minion - BGN 8.20.
- 2 lv.
If the ordered toys are 50 or more the shop makes a discount of 25% from the total price. Petya has to give 10% of the earned money to the shop rent. Calculate whether the money will be enough for her to go on a trip.
Input:
6 lines are read from the console:
1.	Trip price - real number in the range [1.00 ... 10000.00]
2.	Number of puzzles - integer in the range [0 ... 1000]
3.	Number of talking dolls - integer in the range [0 ... 1000]
4.	Number of teddy bears - integer in the range [0 ... 1000]
5.	Number of minions - integer in the range [0 ... 1000]
6.	Number of trucks - integer in the range [0 ... 1000]
Output:
On the console is printed:
- If the money is enough is printed:
o "Yes! {remaining money} lv left."
- If the money is NOT enough is printed:
o "Not enough money! {not enough money} lv needed."
The result should be formatted to the second decimal place.
Example:
Input:         Output:                                                    Input:        Output:
40.8            Yes! 418.20 lv left.                                   320            Not enough money! 238.73 lv needed.
20                                                                              8
25                                                                              2
30                                                                              5
50                                                                              5
10                                                                              1