Condition
Filming for the highly anticipated Godzilla vs Kong movie has begun. Screenwriter Adam Wingard asks you to write a program that calculates whether the projected funds are enough to make the film. The filming will require a certain number of extras, clothing for each extra and a set.
It is known that:
- The decor for the film is worth 10% of the budget. 
- If there are more than 150 extras, there is a 10% discount on the clothing.
Input
3 lines are read from the console:
Line 1.	Budget for the film - a real number in the range [1.00 ... 1000000.00]
Line 2.	Number of extras - integer in the range [1 ... 500]
Line 3.	Cost of clothing per extra - real number in the range [1.00 ... 1000.00]
Output:
Two lines should be printed on the console:
- If the money for the decor and clothing is more than the budget:
o "Not enough money!"
o "Wingard needs {not enough money for the movie} leva more."
- If the money for the set and clothes is less than or equal to the budget:
o "Action!" 
o "Wingard starts filming with {the money left} leva left."
The result must be formatted to the second decimal place.
Example:
Input:         Output:                                                                                                 Input:           Output:
20000        Action!                                                                                                  9587.88        Not enough money!
120            Wingard starts filming with 11340.00 leva left.                                    222               Wingard needs 2495.77 leva more.
55.5                                                                                                                         55.68