Condition
Ivan decides to improve the World Record in long distance swimming. On the console is entered the record in seconds that Ivan has to improve, the distance in meters that he has to swim and the time in seconds for which he swims a distance of 1 m. Write a program that calculates whether he has completed the task, given that: the resistance of the water slows him down every 15 m by 12.5 seconds. When calculating how many times Ivan will slow down as a result of the water resistance, the result should be rounded down to the nearest whole number.
Calculate the time in seconds it will take Ivan to swim the distance and the difference to the World Record. 
Input:
3 lines are read from the console:
1.	The record in seconds - a real number in the range [0.00 ... 100000.00]
2.	Distance in meters - a real number in the range [0.00 ... 100000.00]
3.	The time in seconds to swim a distance of 1 m. - real number in the range [0.00 ... 1000.00]
Output:
The console printout depends on the result:
- If Ivan has improved the World Record (his time is less than the record) we print:
o "Yes, he succeeded! The new world record is { Ivan's time} seconds."
- If he has NOT improved the record (his time is greater than or equal to the record) print:
o "No, he failed! He was {failing seconds} seconds slower."
The result should be formatted to the second decimal place.
Example:
Input:         Output:                                                                                               Input:              Output:
10464        No, he failed! He was 20786.00 seconds slower.                                55555.67         Yes, he succeeded! The new world record is 17688.01 seconds.
1500                                                                                                                      3017
20                                                                                                                          5.03