Condition
Peter wants to buy N video cards, M processors and P number of ram. If the number of video cards is greater than the number of processors he gets a 15% discount on the final bill. The following prices apply:
- Video card - 250 BGN/pc.
- Processor - 35% of the price of purchased video cards/pc.
- RAM - 10% of the price of the purchased video cards/pc.
Calculate the amount needed to purchase the materials and calculate if the budget will be enough.
Input:
The entrance consists of four rows:
1.	Peter's budget - a real number in the range [0.0...100000.0]
2.	The number of video cards - an integer in the range [0...100]
3.	The number of processors - an integer in the range [0...100]
4.	The number of memory frames - integer in the range [0...100]
Output:
On the console, 1 line is printed, which should look like this:
- If the budget is sufficient:
"You have {sufficient budget} left!"
- If the amount exceeds the budget:
"Not enough money! You need {amount needed} leva more!"
Format the result to the second decimal place.
Example:
Input:        Output:                                                          Input:           Output:
900           You have 198.75 leva left!                              920.45          Not enough money! You need 3.92 leva more!
2                                                                                      3
1                                                                                      1
3                                                                                      1