Condition
A pool with a volume of V has two pipes from which it is filled. Each pipe has a certain flow rate (the liters of water passing through a pipe in one hour). The worker runs the pipes in and out simultaneously for N hours. Write a program that outputs the state of the pool at the time the worker returns. 
Input:
Four lines are read from the console:
- The first line contains the number V - Volume of the pool in liters - an integer in the range [1...10000].
- The second line contains the number P1 - flow rate of the first pipe per hour - integer in the range [1...5000].
- The third line contains the number P2 - flow rate of the second pipe per hour - integer in the range [1...5000].
- The fourth line contains the number H - hours the worker is absent - real number in the range [1.0...24.00]
Output:
Print one of the two possible states on the console:
- How much the pool has filled and which pipe contributed what percentage.
o "The pool is {percent full}% full. Pipe 1: {percent water from first pipe}%. Pipe 2: {percent water from second pipe}%."
If the pool overflowed, by how many liters did it overflow in the given time.
o "For {the number of hours the pipes filled with water} hours the pool overflowed with {the number of liters of water in more} liters."
Example:
Input:        Output:                                                                                                         Input:        Output:
1000         The pool is 66.00% full. Pipe 1: 45.45%. Pipe 2: 54.55%.                              100            For 2.50 hours the pool overflows with 400.00 liters.
100                                                                                                                                 100
120                                                                                                                                 100
3                                                                                                                                     2.5