Condition
From a vineyard of X square meters, 40% of the harvest is set aside for wine production. Y kilograms of grapes are produced from 1 square meter of vineyard. For 1 litre of wine, 2.5 kg of grapes are needed. The desired quantity of wine to sell is Z litres.
Write a program that calculates how much wine can be produced and whether this quantity is sufficient. If it is enough, the remainder is divided equally among the workers in the vineyard.
Input:
The input is read from the console and consists of exactly 4 lines:
- 1st line: X sqm is the vineyard - an integer in the range [10 ... 5000]
- 2nd line: Y grapes per sq.m - real number in the range [0.00 ... 10.00]
- 3rd row: Z litres of wine needed - integer in the range [10 ... 600]
- 4th row: number of workers - integer in the range [1 ... 20]
Output:
The following should be printed on the console:
- If the wine produced is less than required:
o "It will be a tough winter! More {not enough wine} liters wine needed."
The result should be rounded to a lower whole number
- If the wine produced is as much or more than needed:
o "Good harvest this year! Total wine: {total wine} liters."
The result should be rounded to a lower whole number
o "{Wine} liters left -> {wine per worker} liters per person."
Both results should be rounded to the higher whole number
Example:
Input:        Output:                                                                                            Input:        Output:
650           Good harvest this year! Total wine: 208 liters.                                 1020          It will be a tough winter! More 180 liters wine needed.
2               33 liters left -> 11 liters per person.                                                1.5
175                                                                                                                    425
3                                                                                                                        4