Condition
Write a program that guesses whether a vehicle's tank needs refueling or not. Two lines are read from the console - text and a real number, the first line reads the fuel type - text with options "Diesel", "Gasoline" or "Gas", and the second reads the liters of fuel that are in the tank. If the litres of fuel are greater than or equal to 25, the console should print "You have enough {fuel type}.", if less than 25, it should print "Fill your tank with {fuel type}!". If fuel other than that specified is entered, print "Invalid fuel!".
Example:
Input:         Output:                                                                              Input:                   Output:
Diesel         Fill your tank with diesel!                                                  Gasoline               You have enough gasoline.
10                                                                                                        40