Condition
Write a program that reads the day of the week (text) in English - entered by the user. If the day is a working day prints on the console - "Working day", if it is a day off - "Weekend". If text other than day of the week is entered it prints - "Error".
Example:
Input:            Output:                                       Input:           Output:                                          Input:          Output:
Monday        Working day                               Sunday        Weekend                                        April            Error