Condition
In one cinema room the chairs are arranged in a rectangular shape in r rows and c columns. There are three types of screenings with tickets at different prices:
- Premiere - premiere screening, priced at 12.00 BGN.
- Normal - standard screening, price 7.50 BGN.
- Discount - screening for children, schoolchildren and students at a reduced price of BGN 5.00.
Write a program that reads the type of screening (string), number of rows and number of columns in the auditorium (integers) entered by the user, and calculates the total ticket revenue when the auditorium is full. Print the result in the format as in the examples below, with 2 decimal places. 
Example:
Input:              Output:                                               Input:           Output:
Premiere         1440.00 leva                                       Normal         2047.50 leva
10                                                                               21
12                                                                               13