Condition
Tony and friends loved to go fishing, they were so keen on fishing that they decided to go fishing by boat. The price for renting the boat depends on the season and the number of fishermen.
The price depends on the season:
- The price for renting the ship in spring is 3000 BGN.
- The price for renting the ship in summer and autumn is 4200 BGN.
- The price for renting the vessel in winter is 2600 BGN.
Depending on the number of your group, a discount is available:
- If the group is up to and including 6 people - 10% discount.
- If the group is from 7 to 11 persons inclusive - discount of 15%.
- If the group is from 12 upwards - discount of 25%. 
Fishermen get an extra 5% discount if they are an even number unless it is autumn - then no extra discount. 
Write a program to calculate if the fishermen will raise enough money. 
Input:
Exactly three lines are read from the console.
- Group budget - integer in the range [1...8000]
- Season - text : "Spring", "Summer", "Autumn", "Winter"
- Number of fishermen - integer in the range [4...18]
Output:
Print one line on the console:
- If the budget is sufficient:
"Yes! You have {remaining money} left."
- If the budget is NOT sufficient:
"Not enough money! You need {amount not enough} leva."
Amounts must be formatted to two decimal places.
Example:
Input:             Output:
3000              Not enough money! You need 570.00 leva.
Summer
11